Add files via upload
Seaborn is a Python module for creating statistical visuals. It is built on top of matplotlib and tightly interacts with pandas' data structures. Seaborn assists you in exploring and comprehending your data.
NumPy is a Python library for manipulating arrays. It also includes functions for working with linear algebra, Fourier transforms, and matrices.
Pandas is a popular open-source Python library for data science/data analysis and machine learning activities. It is constructed on top of Numpy, another library that supports multidimensional arrays.

Exploratory Data Analysis (EDA) is a method of analysing data that employs visual tools. It is used to detect trends and patterns, as well as to test assumptions, using statistical summaries and graphical representations.
Pairplot visualises data to determine the connection between variables that might be continuous or categorical.
The hue option specifies which column in the data frame should be used for colour encoding.